services/auth_service/requirements.txt
flask
flask-sqlalchemy
flask-migrate
psycopg2-binary
PyJWT
werkzeug
python-dotenv
